shared libs: use static event array instead of using static variables for indexing
Using static variable in templates and with shared libraries doesn't play
well.

The static variables counter and value in `Emitter::next_type` and
`Emitter::event_type` are duplicated between the user of the shared library
and the shared library itself.
This leads to wrong handler execution when an event is published.

As the static variables are in a template, there is no easy way to fix
this. One solution is to explicitly instantiate all required instance
of the static variables, but there are hundreds of them.

Instead, the solution here is to not rely on these static variable.
To do this, all possible event types are added to the Emitter class
as a variadic template parameter (`Events`) and the index of an event is
inferred from the Event position in the parameter pack Events.

This has the benefit to check at compile time that the user won't listen
for a wrong event on an handle, for example `ListenEvent` on `AsyncHandle`.

Note: The test `signal.cpp` is registering an event handler for `CheckEvent` on a
`SignalHandle`. As `SignalHandle` doesn't publish any `CheckEvent` it seems,
I've commented the code doing that.
Maybe that's an error and it should be `SignalEvent` instead ?